extract util functions to folder
