+++ server/utils/generateSignedUrl.js
