+++ server/utils/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  generatePreSignedPutUrl: require('./generateSignedUrl'),
+  checkJwt: require('./checkJwt'),
+}
